test: Move geometry tests to separate file
